refator: commonjs -> ECMAScript
